+++ modal/volume.py
@@ -3,6 +3,7 @@ import asyncio
 import concurrent.futures
 import enum
 import functools
+import multiprocessing
 import os
 import platform
 import re
@@ -13,7 +14,6 @@ from dataclasses import dataclass
 from io import BytesIO
 from pathlib import Path, PurePosixPath
 from typing import (
-    IO,
     Any,
     Awaitable,
     BinaryIO,
@@ -32,7 +32,14 @@ from modal_proto import api_pb2
 
 from ._object import EPHEMERAL_OBJECT_HEARTBEAT_SLEEP, _get_environment_name, _Object, live_method, live_method_gen
 from ._resolver import Resolver
-from ._utils.async_utils import TaskContext, aclosing, async_map, asyncnullcontext, synchronize_api
+from ._utils.async_utils import (
+    TaskContext,
+    aclosing,
+    async_map,
+    async_map_ordered,
+    asyncnullcontext,
+    synchronize_api,
+)
 from ._utils.blob_utils import (
     BLOCK_SIZE,
     FileUploadSpec,
@@ -193,6 +200,16 @@ class _Volume(_Object, type_prefix="vo"):
 
     def _get_metadata(self) -> Optional[Message]:
         return self._metadata
+
+
+    @property
+    def _is_v1(self) -> bool:
+        return self._metadata.version in [
+            None,
+            api_pb2.VolumeFsVersion.VOLUME_FS_VERSION_UNSPECIFIED,
+            api_pb2.VolumeFsVersion.VOLUME_FS_VERSION_V1
+        ]
+
 
     @classmethod
     @asynccontextmanager
@@ -390,7 +407,7 @@ class _Volume(_Object, type_prefix="vo"):
         return [entry async for entry in self.iterdir(path, recursive=recursive)]
 
     @live_method_gen
-    async def read_file(self, path: str) -> AsyncIterator[bytes]:
+    def read_file(self, path: str) -> AsyncIterator[bytes]:
         """
         Read a file from the modal.Volume.
 
@@ -404,6 +421,10 @@ class _Volume(_Object, type_prefix="vo"):
         print(len(data))  # == 1024 * 1024
         ```
         """
+        return self._read_file1(path) if self._is_v1 else self._read_file2(path)
+
+
+    async def _read_file1(self, path: str) -> AsyncIterator[bytes]:
         req = api_pb2.VolumeGetFileRequest(volume_id=self.object_id, path=path)
         try:
             response = await retry_transient_errors(self._client.stub.VolumeGetFile, req)
@@ -417,53 +438,30 @@ class _Volume(_Object, type_prefix="vo"):
             async for data in blob_iter(response.data_blob_id, self._client.stub):
                 yield data
 
-    @live_method
-    async def read_file_into_fileobj(self, path: str, fileobj: IO[bytes]) -> int:
-        """mdmd:hidden
 
-        Read volume file into file-like IO object.
-        In the future, this will replace the current generator implementation of the `read_file` method.
-        """
+    async def _read_file2(self, path: str) -> AsyncIterator[bytes]:
+        req = api_pb2.VolumeGetFile2Request(volume_id=self.object_id, path=path)
 
-        chunk_size_bytes = 8 * 1024 * 1024
-        start = 0
-        req = api_pb2.VolumeGetFileRequest(volume_id=self.object_id, path=path, start=start, len=chunk_size_bytes)
         try:
-            response = await retry_transient_errors(self._client.stub.VolumeGetFile, req)
+            response = await retry_transient_errors(self._client.stub.VolumeGetFile2, req)
         except GRPCError as exc:
             raise FileNotFoundError(exc.message) if exc.status == Status.NOT_FOUND else exc
-        if response.WhichOneof("data_oneof") != "data":
-            raise RuntimeError("expected to receive 'data' in response")
 
-        n = fileobj.write(response.data)
-        if n != len(response.data):
-            raise OSError(f"failed to write {len(response.data)} bytes to output. Wrote {n}.")
-        elif n == response.size:
-            return response.size
-        elif n > response.size:
-            raise RuntimeError(f"length of returned data exceeds reported filesize: {n} > {response.size}")
-        # else: there's more data to read. continue reading with further ranged GET requests.
-        file_size = response.size
-        written = n
+        async def read_block(block_url: str) -> bytes:
+            async with ClientSessionRegistry.get_session().get(block_url) as get_response:
+                return await get_response.content.read()
 
-        while True:
-            req = api_pb2.VolumeGetFileRequest(volume_id=self.object_id, path=path, start=written, len=chunk_size_bytes)
-            response = await retry_transient_errors(self._client.stub.VolumeGetFile, req)
-            if response.WhichOneof("data_oneof") != "data":
-                raise RuntimeError("expected to receive 'data' in response")
-            if len(response.data) > chunk_size_bytes:
-                raise RuntimeError(f"received more data than requested: {len(response.data)} > {chunk_size_bytes}")
-            elif (written + len(response.data)) > file_size:
-                raise RuntimeError(f"received data exceeds filesize of {chunk_size_bytes}")
+        async def iter_urls() -> AsyncGenerator[str]:
+            for url in response.get_urls:
+                yield url
 
-            n = fileobj.write(response.data)
-            if n != len(response.data):
-                raise OSError(f"failed to write {len(response.data)} bytes to output. Wrote {n}.")
-            written += n
-            if written == file_size:
-                break
+        # TODO(dflemstr): Reasonable default? Make configurable?
+        prefetch_num_blocks = multiprocessing.cpu_count()
 
-        return written
+        async with aclosing(async_map_ordered(iter_urls(), read_block, concurrency=prefetch_num_blocks)) as stream:
+            async for value in stream:
+                yield value
+
 
     @live_method
     async def remove_file(self, path: str, recursive: bool = False) -> None:
@@ -755,19 +753,27 @@ class _VolumeUploadContextManager(_AbstractVolumeUploadContextManager):
 
 VolumeUploadContextManager = synchronize_api(_VolumeUploadContextManager)
 
-_FileUploader2 = Callable[[], Awaitable[FileUploadSpec2]]
+_FileUploader2 = Callable[[asyncio.Semaphore], Awaitable[FileUploadSpec2]]
 
 class _VolumeUploadContextManager2(_AbstractVolumeUploadContextManager):
     """Context manager for batch-uploading files to a Volume version 2."""
 
     _volume_id: str
     _client: _Client
-    _force: bool
     _progress_cb: Callable[..., Any]
+    _force: bool
+    _hash_concurrency: int
+    _put_concurrency: int
     _uploader_generators: list[Generator[_FileUploader2]]
 
     def __init__(
-        self, volume_id: str, client: _Client, progress_cb: Optional[Callable[..., Any]] = None, force: bool = False
+        self,
+        volume_id: str,
+        client: _Client,
+        progress_cb: Optional[Callable[..., Any]] = None,
+        force: bool = False,
+        hash_concurrency: int = multiprocessing.cpu_count(),
+        put_concurrency: int = multiprocessing.cpu_count(),
     ):
         """mdmd:hidden"""
         self._volume_id = volume_id
@@ -775,6 +781,8 @@ class _VolumeUploadContextManager2(_AbstractVolumeUploadContextManager):
         self._uploader_generators = []
         self._progress_cb = progress_cb or (lambda *_, **__: None)
         self._force = force
+        self._hash_concurrency = hash_concurrency
+        self._put_concurrency = put_concurrency
 
     async def __aenter__(self):
         return self
@@ -787,7 +795,9 @@ class _VolumeUploadContextManager2(_AbstractVolumeUploadContextManager):
                     yield from gen
 
             async def gen_file_upload_specs() -> list[FileUploadSpec2]:
-                uploads = [asyncio.create_task(fut()) for fut in gen_upload_providers()]
+                hash_semaphore = asyncio.Semaphore(self._hash_concurrency)
+
+                uploads = [asyncio.create_task(fut(hash_semaphore)) for fut in gen_upload_providers()]
                 logger.debug(f"Computing checksums for {len(uploads)} files")
 
                 file_specs = []
@@ -817,9 +827,19 @@ class _VolumeUploadContextManager2(_AbstractVolumeUploadContextManager):
 
         def gen():
             if isinstance(local_file, str) or isinstance(local_file, Path):
-                yield lambda: FileUploadSpec2.from_path(local_file, PurePosixPath(remote_path), mode)
+                yield lambda hash_semaphore: FileUploadSpec2.from_path(
+                    local_file,
+                    PurePosixPath(remote_path),
+                    hash_semaphore,
+                    mode
+                )
             else:
-                yield lambda: FileUploadSpec2.from_fileobj(local_file, PurePosixPath(remote_path), mode or 0o644)
+                yield lambda hash_semaphore: FileUploadSpec2.from_fileobj(
+                    local_file,
+                    PurePosixPath(remote_path),
+                    hash_semaphore,
+                    mode or 0o644
+                )
 
         self._uploader_generators.append(gen())
 
@@ -840,7 +860,7 @@ class _VolumeUploadContextManager2(_AbstractVolumeUploadContextManager):
 
         def create_spec(subpath):
             relpath_str = subpath.relative_to(local_path)
-            return lambda: FileUploadSpec2.from_path(subpath, remote_path / relpath_str)
+            return lambda hash_semaphore: FileUploadSpec2.from_path(subpath, remote_path / relpath_str, hash_semaphore)
 
         def gen():
             glob = local_path.rglob("*") if recursive else local_path.glob("*")
@@ -854,6 +874,8 @@ class _VolumeUploadContextManager2(_AbstractVolumeUploadContextManager):
     async def _put_file_specs(self, file_specs: list[FileUploadSpec2]):
         put_responses = {}
         # num_blocks_total = sum(len(file_spec.blocks_sha256) for file_spec in file_specs)
+
+        logger.debug(f"Ensuring {len(file_specs)} files are uploaded...")
 
         # We should only need two iterations: Once to possibly get some missing_blocks; the second time we should have
         # all blocks uploaded
@@ -888,7 +910,13 @@ class _VolumeUploadContextManager2(_AbstractVolumeUploadContextManager):
             if not response.missing_blocks:
                 break
 
-            await _put_missing_blocks(file_specs, response.missing_blocks, put_responses, self._progress_cb)
+            await _put_missing_blocks(
+                file_specs,
+                response.missing_blocks,
+                put_responses,
+                self._put_concurrency,
+                self._progress_cb
+            )
         else:
             raise RuntimeError("Did not succeed at uploading all files despite supplying all missing blocks")
 
@@ -904,8 +932,19 @@ async def _put_missing_blocks(
     # by the nested class (?)
     missing_blocks: list,
     put_responses: dict[bytes, bytes],
+    put_concurrency: int,
     progress_cb: Callable[..., Any]
 ):
+    @dataclass
+    class FileProgress:
+        task_id: str
+        pending_blocks: set[int]
+
+    put_semaphore = asyncio.Semaphore(put_concurrency)
+    file_progresses: dict[str, FileProgress] = dict()
+
+    logger.debug(f"Uploading {len(missing_blocks)} missing blocks...")
+
     async def put_missing_block(
         # TODO(dflemstr): Type is `api_pb2.VolumePutFiles2Response.MissingBlock` but synchronicity gets confused
         # by the nested class (?)
@@ -923,23 +962,36 @@ async def _put_missing_blocks(
         block_start = missing_block.block_index * BLOCK_SIZE
         block_length = min(BLOCK_SIZE, file_spec.size - block_start)
 
-        progress_name = f"{file_spec.path} block {missing_block.block_index + 1} / {len(file_spec.blocks_sha256)}"
-        progress_task_id = progress_cb(name=progress_name, size=file_spec.size)
+        if file_spec.path not in file_progresses:
+            file_task_id = progress_cb(name=file_spec.path, size=file_spec.size)
+            file_progresses[file_spec.path] = FileProgress(task_id=file_task_id, pending_blocks=set())
 
-        with file_spec.source() as source_fp:
-            payload = BytesIOSegmentPayload(
-                source_fp,
-                block_start,
-                block_length,
-                progress_report_cb=functools.partial(progress_cb, progress_task_id)
-            )
+        file_progress = file_progresses[file_spec.path]
+        file_progress.pending_blocks.add(missing_block.block_index)
+        task_progress_cb = functools.partial(progress_cb, task_id=file_progress.task_id)
 
-            async with ClientSessionRegistry.get_session().put(
-                missing_block.put_url,
-                data=payload,
-            ) as response:
-                response.raise_for_status()
-                resp_data = await response.content.read()
+        async with put_semaphore:
+            with file_spec.source() as source_fp:
+                payload = BytesIOSegmentPayload(
+                    source_fp,
+                    block_start,
+                    block_length,
+                    # limit chunk size somewhat to not keep event loop busy for too long
+                    chunk_size=256*1024,
+                    progress_report_cb=task_progress_cb
+                )
+
+                async with ClientSessionRegistry.get_session().put(
+                    missing_block.put_url,
+                    data=payload,
+                ) as response:
+                    response.raise_for_status()
+                    resp_data = await response.content.read()
+
+        file_progress.pending_blocks.remove(missing_block.block_index)
+
+        if len(file_progress.pending_blocks) == 0:
+            task_progress_cb(complete=True)
 
         return block_sha256, resp_data