+++ test/conftest.py
@@ -69,14 +69,11 @@ class VolumeFile:
     data_blob_id: Optional[str] = None
     block_hashes: list[bytes] = dataclasses.field(default_factory=list)
 
-
 @dataclasses.dataclass
 class GrpcErrorAndCount:
-    """Helper class that holds a gRPC error and the number of times it should be raised."""
-
+    """ Helper class that holds a gRPC error and the number of times it should be raised. """
     grpc_error: Status
     count: int
-
 
 # TODO: Isolate all test config from the host
 @pytest.fixture(scope="function", autouse=True)
@@ -1302,13 +1299,21 @@ class MockClientServicer(api_grpc.ModalClientBase):
 
     async def FunctionUpdateSchedulingParams(self, stream):
         req: api_pb2.FunctionUpdateSchedulingParamsRequest = await stream.recv_message()
+
         # update function definition
         fn_definition = self.app_functions[req.function_id]
         assert isinstance(fn_definition, api_pb2.Function)
-        # Hacky that we're modifying the function definition directly
-        # In the server we track autoscaler updates separately
+        # Note that this doesn't mock the full server logic very well
         fn_definition.warm_pool_size = req.warm_pool_size_override
         fn_definition.autoscaler_settings.MergeFrom(req.settings)
+
+        # Hacky that we're modifying the function definition directly
+        # In the server we track autoscaler updates separately
+        fn_definition.warm_pool_size = fn_definition.autoscaler_settings.min_containers
+        fn_definition.concurrency_limit = fn_definition.autoscaler_settings.max_containers
+        fn_definition._experimental_buffer_containers = fn_definition.autoscaler_settings.buffer_containers
+        fn_definition.task_idle_timeout_secs = fn_definition.autoscaler_settings.scaledown_window
+
         await stream.send_message(api_pb2.FunctionUpdateSchedulingParamsResponse())
 
     ### Image
@@ -1882,6 +1887,39 @@ class MockClientServicer(api_grpc.ModalClientBase):
             else:
                 await stream.send_message(api_pb2.VolumeGetFileResponse(data=vol_file.data, size=size))
 
+    async def VolumeGetFile2(self, stream):
+        req = await stream.recv_message()
+        if req.path not in self.volumes[req.volume_id].files:
+            raise GRPCError(Status.NOT_FOUND, "File not found")
+        vol_file = self.volumes[req.volume_id].files[req.path]
+        get_urls = []
+
+        def ceildiv(a: int, b: int) -> int:
+            return -(a // -b)
+
+        total_start = req.start
+        total_end = req.start + (req.len or len(vol_file.data))
+
+        block_start = min(total_start // BLOCK_SIZE, len(vol_file.block_hashes))
+        block_end = min(ceildiv(total_end, BLOCK_SIZE), len(vol_file.block_hashes))
+
+        for idx, block_hash in enumerate(vol_file.block_hashes[block_start:block_end]):
+            # Crude port of internal `blocks_for_byte_slice` algorithm:
+            start = total_start % BLOCK_SIZE if idx == 0 else 0
+            end = (((total_end - 1) % BLOCK_SIZE) + 1 if total_end > 0 else 0) \
+                if idx == (block_end - block_start - 1) else BLOCK_SIZE
+            length = end - start
+
+            get_urls.append(f"{self.blob_host}/block/test-get-request:{block_hash.hex()}:{start}:{length}")
+
+        response = api_pb2.VolumeGetFile2Response(
+            get_urls=get_urls,
+            size=len(vol_file.data),
+            start=total_start,
+            len=total_end - total_start
+        )
+        await stream.send_message(response)
+
     async def VolumeRemoveFile(self, stream):
         req = await stream.recv_message()
         if req.path not in self.volumes[req.volume_id].files:
@@ -1983,7 +2021,7 @@ class MockClientServicer(api_grpc.ModalClientBase):
                     missing_block = api_pb2.VolumePutFiles2Response.MissingBlock(
                         file_index=file_index,
                         block_index=block_index,
-                        put_url=f"{self.blob_host}/block?token=test-put-request",
+                        put_url=f"{self.blob_host}/block/test-put-request",
                     )
                     file_missing_blocks.append(missing_block)
 
@@ -2062,7 +2100,7 @@ def blob_server():
         return aiohttp.web.Response(body=blobs[blob_id])
 
     async def put_block(request):
-        token = request.query["token"]
+        token = request.match_info["token"]
         if token != "test-put-request":
             return aiohttp.web.Response(status=400, text="bad token")
 
@@ -2077,13 +2115,26 @@ def blob_server():
         blocks[block_id] = content
         return aiohttp.web.Response(text=f"test-put-response:{block_id}")
 
+    async def get_block(request):
+        token = request.match_info["token"]
+
+        magic, block_id, start, length = token.split(':')
+        if magic != "test-get-request":
+            return aiohttp.web.Response(status=400, text="bad token")
+
+        start = int(start)
+        length = int(length)
+
+        return aiohttp.web.Response(body=blocks[block_id][start:start+length])
+
     app = aiohttp.web.Application()
     app.add_routes([aiohttp.web.put("/upload", upload)])
     app.add_routes([aiohttp.web.get("/download", download)])
     app.add_routes([aiohttp.web.post("/complete_multipart", complete_multipart)])
 
     # API used for volume version 2 blocks:
-    app.add_routes([aiohttp.web.put("/block", put_block)])
+    app.add_routes([aiohttp.web.get("/block/{token}", get_block)])
+    app.add_routes([aiohttp.web.put("/block/{token}", put_block)])
 
     started = threading.Event()
     stop_server = threading.Event()